requirements.txt
pdfplumber
sentence-transformers
faiss-cpu
langchain
langchain-openai
openai
tqdm
keybert
unstructured[local-inference,pdf] 